Réparation dernier push

+++ XBee/include/xbee.h
@@ -17,6 +17,7 @@
 #include <functional>
 #include <robotech/logs.h>
 
+#include "serialib.h"
 #include "define_xbee.h"
 
 
@@ -76,7 +77,7 @@ private:
     int processBuffer(const std::vector<uint8_t> &response);
     int processFrame(const std::vector<uint8_t> &buffer);
     template<typename T> void readRx(T &buffer, unsigned int timeout = 100);
-    static uint16_t computeChecksum(const std::vector<uint8_t> &frame, uint8_t start, uint8_t length);
+    static uint8_t computeChecksum(const std::vector<uint8_t> &frame, uint8_t start, uint8_t length);
 };
